backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]>=3.2.10
dj-database-url==2.3.0
python-decouple==3.8
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
